+++ src/core_plugins/kibana/public/context/app.js
@@ -3,7 +3,7 @@ import _ from 'lodash';
 import uiModules from 'ui/modules';
 import contextAppTemplate from './app.html';
 import './components/loading_button';
-import './components/size_picker';
+import './components/size_picker/size_picker';
 import { bindSelectors } from './utils/selectors';
 import {
   QueryParameterActionsProvider,
