+++ src/core_plugins/kibana/public/context/components/size_picker/size_picker.html
@@ -1,16 +1,6 @@
-<div class="contextSizePicker">
-  <input
-    class="kuiTextInput contextSizePicker__item contextSizePicker__sizeInput"
-    ng-model-options="{updateOn: 'change', getterSetter: true, debounce: 200}"
-    ng-model="vm.getOrSetPredecessorCount"
-    type="number"
-  />
-  <div class="contextSizePicker__item">newer and</div>
-  <input
-    class="kuiTextInput contextSizePicker__item contextSizePicker__sizeInput"
-    ng-model-options="{updateOn: 'change', getterSetter: true, debounce: 200}"
-    ng-model="vm.getOrSetSuccessorCount"
-    type="number"
-  />
-  <div class="contextSizePicker__item">older entries</div>
-</div>
+<input
+  class="kuiTextInput contextSizePicker"
+  ng-model-options="{updateOn: 'change', getterSetter: true, debounce: 200}"
+  ng-model="model"
+  type="number"
+>
