+++ src/core_plugins/kibana/public/context/components/size_picker/size_picker.less
@@ -1,20 +1,14 @@
+/**
+ * 1. Hide increment and decrement buttons for type="number" input.
+ */
 .contextSizePicker {
-  padding-right: 10px;
-  height: 100%;
+  appearance: textfield;
+  text-align: center;
+  width: 3em;
+
+  &::-webkit-outer-spin-button,
+  &::-webkit-inner-spin-button {
+    appearance: none; /* 1 */
+    margin: 0; /* 1 */
+  }
 }
-
-  .contextSizePicker__item {
-    display: inline-block;
-  }
-
-  .contextSizePicker__sizeInput {
-    appearance: textfield;
-    text-align: center;
-    width: 3em;
-
-    &::-webkit-outer-spin-button,
-    &::-webkit-inner-spin-button {
-      appearance: none;
-      margin: 0;
-    }
-  }
