+++ src/ui/public/styles/base.less
@@ -806,15 +806,6 @@ fieldset {
   color: #F39C12;
 }
 
-.kuiInlineFormGroup {
-  display: flex;
-  align-items: center;
-
-  > * + * {
-    margin-left: 5px;
-  }
-}
-
 /**
  * 1. Override Bootstrap ui-select component styles. If side padding is > 5px, the component breaks.
  */
