+++ src/ui_framework/components/bar/_bar.scss
@@ -1,0 +1,7 @@
+.kuiBar {
+  @include bar;
+}
+
+.kuiBarSection {
+  @include barSection;
+}
