+++ src/ui_framework/components/bar/_index.scss
@@ -1,0 +1,1 @@
+@import "bar";
