+++ src/ui_framework/doc_site/src/views/bar/bar.html
@@ -1,0 +1,18 @@
+ <div class="kuiBar">
+  <div class="kuiBarSection">
+    <div class="kuiButtonGroup">
+      <button class="kuiButton kuiButton--basic">
+        See previous 10 pages
+      </button>
+      <button class="kuiButton kuiButton--basic">
+        See next 10 pages
+      </button>
+    </div>
+  </div>
+
+  <div class="kuiBarSection">
+    <div>Limit to</div>
+    <input class="kuiTextInput" size="2" value="10">
+    <div>pages</div>
+  </div>
+</div>
