+++ src/ui_framework/doc_site/src/views/bar/bar_one_section.html
@@ -1,0 +1,12 @@
+ <div class="kuiBar">
+  <div class="kuiBarSection">
+    <div class="kuiButtonGroup">
+      <button class="kuiButton kuiButton--basic">
+        See previous 10 pages
+      </button>
+      <button class="kuiButton kuiButton--basic">
+        See next 10 pages
+      </button>
+    </div>
+  </div>
+</div>
