+++ src/ui_framework/doc_site/src/views/bar/bar_three_sections.html
@@ -1,0 +1,38 @@
+ <div class="kuiBar">
+  <div class="kuiBarSection">
+    <div class="kuiButtonGroup">
+      <button class="kuiButton kuiButton--basic">
+        See previous 10 pages
+      </button>
+      <button class="kuiButton kuiButton--basic">
+        See next 10 pages
+      </button>
+    </div>
+  </div>
+
+  <div class="kuiBarSection">
+    <div class="kuiButtonGroup">
+      <button class="kuiButton kuiButton--basic">
+        Create new page
+      </button>
+      <button class="kuiButton kuiButton--danger">
+        Clear all pages
+      </button>
+    </div>
+  </div>
+
+  <div class="kuiBarSection">
+    <div>Limit to</div>
+    <input class="kuiTextInput" size="2" value="10">
+    <div>pages</div>
+
+    <div class="kuiButtonGroup">
+      <button class="kuiButton kuiButton--basic">
+        Undo
+      </button>
+      <button class="kuiButton kuiButton--basic">
+        Redo
+      </button>
+    </div>
+  </div>
+</div>
