+++ ui_framework/components/index.scss
@@ -44,6 +44,10 @@ $tableBorder: 2px solid $panelColor;
 // Timing
 $formTransitionTiming: 0.1s linear;
 
+// Bar
+$toolBarSectionSpacing: 50px;
+$toolBarItsemSpacing: 10px;
+
 @mixin darkTheme() {
   .theme-dark & {
     @content;
@@ -94,6 +98,40 @@ $formTransitionTiming: 0.1s linear;
   }
 }
 
+@mixin bar {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+}
+
+/**
+ * 1. Put 10px of space between each child.
+ * 2. If there is only one child, align it to the right. If you wanted it aligned right, you
+ *    wouldn't use the Bar in the first place.
+ */
+@mixin barSection {
+  display: flex;
+  align-items: center;
+  margin-left: $toolBarSectionSpacing * 0.5;
+  margin-right: $toolBarSectionSpacing * 0.5;
+
+  &:first-child {
+    margin-left: 0;
+  }
+
+  &:last-child {
+    margin-right: 0;
+  }
+
+  &:only-child {
+    margin-left: auto; /* 1 */
+  }
+
+  & > * + * {
+    margin-left: $toolBarItsemSpacing; /* 1 */
+  }
+}
+
 * {
   box-sizing: border-box;
 }
@@ -102,6 +140,7 @@ body {
   font-family: $font;
 }
 
+@import "bar/index";
 @import "button/index";
 @import "form/index";
 @import "icon/index";
