+++ ui_framework/components/tool_bar/_index.scss
@@ -1,6 +1,3 @@
-$toolBarSectionSpacing: 50px;
-$toolBarItsemSpacing: 10px;
-
 @import "tool_bar";
 @import "tool_bar_footer";
 @import "tool_bar_search";
