+++ ui_framework/components/tool_bar/_tool_bar.scss
@@ -1,27 +1,12 @@
 .kuiToolBar {
-  display: flex;
-  justify-content: space-between;
+  @include bar;
+  @include buttonOnReverseBackground;
+
   padding: 10px;
   height: 50px;
   background-color: $panelColor;
-
-  @include buttonOnReverseBackground;
 }
 
-/**
- * 1. Put 10px of space between each child.
- */
 .kuiToolBarSection {
-  display: flex;
-  align-items: center;
-  margin-left: $toolBarSectionSpacing * 0.5;
-  margin-right: $toolBarSectionSpacing * 0.5;
-
-  &:last-child {
-    margin-right: 0;
-  }
-
-  & > * + * {
-    margin-left: $toolBarItsemSpacing; /* 1 */
-  }
+  @include barSection;
 }
